Whole new batch of tests! Every express device route has beed tested simulating multiple situations (including the ones described in the challenge), testing response status, database changes and etc

+++ test/initializer.test.js
@@ -7,23 +7,19 @@ before(async () => {
       // We are using a database only for testing purposes
       await mongoose.connect(config.database.uri);
       console.log('Connected to MongoDB!');
+
+      // Before each test, we get all the collections and drop then
+      // Like this, we ensure no 'ns not found' exception is thrown and keep it safer
+      let colls = await mongoose.connection.db.listCollections().toArray();
+      for (let coll of colls) {
+        if (coll.name !== 'system.indexes'){
+          await mongoose.connection.collections[coll.name].drop();
+        }
+      }
     }
     catch (error) {
       console.log('Connection error : ', error);
     }
-});
- 
-beforeEach(async () => {
-  try{
-    // Before each test, we get all the collections and drop then
-    // Like this, we ensure no 'ns not found' exception is thrown and keep it safer
-    let colls = await mongoose.connection.db.listCollections({name: 'devices'}).toArray();
-    for (let coll of colls) {
-      await mongoose.connection.collections[coll.name].drop();
-    }
-  } catch (error) {
-    console.log('Dropping collection error: ', error);
-  } 
 });
 
 after(async () => {
